An android store product list app.
This app enables users to keep track of their stores' products. The user
can easily create and edit a list of products with their product names,
photos, quantity and price. Also users can track a sale or or reception
of a shipment and in addition they are able to order more products by
email.